renovate: Attempt to not pin node in mise.toml
The only reason node is specified in mise.toml is to ensure that the
other tools run under compatible node version. This is why it's not
desired to pin it in the configuration file.

However, it is locked still in the lock file to ensure CI stability and
renovate should bump that one when there is new version.